Move placeholder lines up a level
