+++ packages/help-center/src/types.ts
@@ -53,7 +53,7 @@ export interface FeatureFlags {
 }
 
 export interface SearchResult {
-	railcar: {
+	railcar?: {
 		railcar?: string;
 		fetch_algo?: string;
 		fetch_lang?: string;
